Moved and renamed a couple of files
