+++ examples/sklearn_learner.py
@@ -6,7 +6,7 @@ from sklearn.utils.validation import check_is_fitted
 from tqdm import tqdm
 
 from config import Config
-from model import BasicLearner, LearnerData
+from basic_learner import BasicLearner, LearnerData
 
 
 class SKLearnWeights:
