+++ examples/xray/data.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from colearn.config import Config
 from examples.utils.data import shuffle_data, split_by_chunksizes
 from examples.xray_utils.data import estimate_cases, train_generator
-from model import LearnerData
+from basic_learner import LearnerData
 
 normal_fl = "normal.pickle"
 pneu_fl = "pneumonia.pickle"
